+++ config/config.go
@@ -8,6 +8,7 @@ type Configuration struct {
 	Jwt       Jwt       `mapstructure:"jwt" json:"jwt" yaml:"jwt"`
 	Redis     Redis     `mapstructure:"redis" json:"redis" yaml:"redis"`
 	RabbitMQ  RabbitMQ  `mapstructure:"rabbitmq" json:"rabbitMQ" yaml:"rabbitMQ"`
+	Kafka     Kafka     `mapstructure:"kafka" json:"kafka" yaml:"kafka"`
 	Cron      Cron      `mapstructure:"cron" json:"cron" yaml:"cron"`
 	WebSocket WebSocket `mapstructure:"websocket" json:"websocket" yaml:"websocket"`
 	ApiUrls   ApiUrls   `mapstructure:"api_url" json:"api_url" yaml:"api_url"`